src/domain/forum/app/use-cases/comment-on-question.ts
import { QuestionsRepository } from '../repositories/questions-repository';
import { QuestionComment } from '../../enterprise/entities/question-comments';
import { QuestionCommentsRepository } from '../repositories/question-comments-repository';
import { UniqueEntityId } from '@/core/domain/value-objects/unique-entity-id';
import { Either, left, right } from '@/core/__error/either';
import { ResourceNotFoundError } from '../../../../core/__error/__errors/resource-not-found-error';
import { Injectable } from '@nestjs/common';

export interface CommentOnQuestionUseCaseRequest {
  authorId: string;
  questionId: string;
  content: string;
}

export type CommentOnQuestionUseCaseResponse = Either<
  ResourceNotFoundError,
  {
    questionComment: QuestionComment;
  }
>;

@Injectable()
export class CommentOnQuestionUseCase {
  constructor(
    private readonly questionsRepository: QuestionsRepository,
    private readonly questionCommentsRepository: QuestionCommentsRepository,
  ) {}
  async execute({
    authorId,
    questionId,
    content,
  }: CommentOnQuestionUseCaseRequest): Promise<CommentOnQuestionUseCaseResponse> {
    const question = await this.questionsRepository.findById(questionId);

    if (!question) {
      return left(new ResourceNotFoundError());
    }

    const newComment = QuestionComment.create({
      authorId: new UniqueEntityId(authorId),
      questionId: new UniqueEntityId(questionId),
      content,
    });

    await this.questionCommentsRepository.create(newComment);

    return right({
      questionComment: newComment,
    });
  }
}
